requirements.txt
amortization==2.4.0
asgiref==3.8.1
Django==4.2.16
et_xmlfile==2.0.0
Jinja2==3.1.4
MarkupSafe==3.0.2
numpy==2.0.2
numpy-financial==1.0.0
openpyxl==3.1.5
pandas==2.2.3
python-dateutil==2.9.0.post0
pytz==2024.2
six==1.16.0
sqlparse==0.5.1
tabulate==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
